@@ -12,69 +12,69 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "6.0.0-beta.7",
-        "@angular/common": "6.0.0-beta.7",
-        "@angular/compiler": "6.0.0-beta.7",
-        "@angular/core": "6.0.0-beta.7",
-        "@angular/forms": "6.0.0-beta.7",
-        "@angular/http": "6.0.0-beta.7",
-        "@angular/platform-browser": "6.0.0-beta.7",
-        "@angular/platform-browser-dynamic": "6.0.0-beta.7",
-        "@angular/router": "6.0.0-beta.7",
-        "@ionic-native/app-version": "4.7.0",
-        "@ionic-native/barcode-scanner": "4.7.0",
-        "@ionic-native/calendar": "4.7.0",
-        "@ionic-native/camera": "4.7.0",
-        "@ionic-native/contacts": "4.7.0",
-        "@ionic-native/core": "4.7.0",
-        "@ionic-native/device": "4.7.0",
-        "@ionic-native/file": "4.7.0",
-        "@ionic-native/file-opener": "4.7.0",
-        "@ionic-native/geolocation": "4.7.0",
-        "@ionic-native/media-capture": "4.7.0",
-        "@ionic-native/network": "4.7.0",
-        "@ionic-native/vibration": "4.7.0",
-        "animate.css": "^3.5.2",
-        "bootstrap": "3.3.7",
+        "@angular/animations": "6.1.0-rc.3",
+        "@angular/common": "6.1.0-rc.3",
+        "@angular/compiler": "6.1.0-rc.3",
+        "@angular/core": "6.1.0-rc.3",
+        "@angular/forms": "6.1.0-rc.3",
+        "@angular/http": "6.1.0-rc.3",
+        "@angular/platform-browser": "6.1.0-rc.3",
+        "@angular/platform-browser-dynamic": "6.1.0-rc.3",
+        "@angular/router": "6.1.0-rc.3",
+        "@ionic-native/app-version": "4.10.0",
+        "@ionic-native/barcode-scanner": "4.10.0",
+        "@ionic-native/calendar": "4.10.0",
+        "@ionic-native/camera": "4.10.0",
+        "@ionic-native/contacts": "4.10.0",
+        "@ionic-native/core": "4.10.0",
+        "@ionic-native/device": "4.10.0",
+        "@ionic-native/file": "4.10.0",
+        "@ionic-native/file-opener": "4.10.0",
+        "@ionic-native/geolocation": "4.10.0",
+        "@ionic-native/media-capture": "4.10.0",
+        "@ionic-native/network": "4.10.0",
+        "@ionic-native/vibration": "4.10.0",
+        "angular2-websocket": "0.9.6",
+        "animate.css": "3.6.1",
+        "bootstrap": "4.1.3",
         "core-js": "2.5.7",
-        "d3": "3.5.17",
+        "d3": "5.5.0",
         "font-awesome": "4.7.0",
-        "fullcalendar": "3.0.1",
+        "fullcalendar": "3.9.0",
         "hammerjs": "2.0.8",
         "iscroll": "5.2.0",
         "jquery": "3.3.1",
         "jquery-ui": "1.12.1",
-        "js-cookie": "^2.2.0",
-        "lodash": "4.17.5",
-        "moment": "2.21.0",
-        "ngx-bootstrap": "2.0.5",
-        "ngx-color-picker": "5.3.0",
-        "ngx-mask": "2.9.6",
-        "ngx-toastr": "8.8.0",
-        "nvd3": "1.8.4",
-        "rxjs": "5.5.2",
+        "js-cookie": "2.2.0",
+        "lodash": "4.17.10",
+        "moment": "2.22.2",
+        "ngx-bootstrap": "3.0.1",
+        "ngx-color-picker": "6.5.0",
+        "ngx-mask": "6.1.2",
+        "ngx-toastr": "8.10.0",
+        "nvd3": "1.8.6",
+        "rxjs": "6.2.2",
+        "rxjs-compat": "6.2.2",
         "summernote": "0.8.10",
-        "webpack": "3.11.0",
+        "webpack": "4.16.2",
         "x2js": "3.2.1",
         "zone.js": "0.8.26"
     },
     "devDependencies": {
-        "@angular/cli": "1.7.3",
-        "@angular/compiler-cli": "6.0.0-beta.4",
-        "@angular/language-service": "6.0.0-beta.4",
-        "angular2-websocket": "^0.9.6",
+        "@angular/cli": "6.1.0-rc.3",
+        "@angular/compiler-cli": "6.1.0-rc.3",
+        "@angular/language-service": "6.1.0-rc.3",
         "@types/jquery": "3.3.1",
-        "@types/node": "10.3.4",
-        "codelyzer": "4.3.0",
+        "@types/node": "10.5.3",
+        "codelyzer": "4.4.2",
         "core-js-builder": "2.5.7",
-        "gulp": "3.9.1",
-        "gulp-angular-embed-templates": "2.3.0",
         "rimraf": "2.6.2",
-        "rollup": "0.60.7",
-        "run-sequence": "2.2.1",
-        "ts-node": "6.1.1",
+        "rollup": "0.63.4",
+        "rollup-plugin-commonjs": "9.1.3",
+        "rollup-plugin-node-resolve": "3.3.0",
+        "ts-node": "7.0.0",
         "tslint": "5.10.0",
-        "typescript": "2.6.2",
-        "uglify-js": "3.4.0"
+        "typescript": "2.9.2",
+        "uglify-js": "3.4.5"
     }
 }
