@@ -1,5 +1,7 @@
 import { Injectable } from '@angular/core';
-import { Subject } from 'rxjs/Subject';
+
+import { Subject } from 'rxjs';
+
 import { AbstractSpinnerService } from '@wm/core';
 
 declare const _;
